feat: Add ad submission functionality
The code sends a POST request to the server at
"http://localhost:3001/advertisements" to submit a user's advertisement
data. The data is sent in a FormData object, including the title,
description, price, userID, username, and images of the ad. The server
response is then checked, and if the status is 200, the user is
navigated to the ads page.